refine

+++ test/ir/pir/cinn/symbolic/test_llama_slice_concat.py
@@ -83,7 +83,7 @@ class SliceConcatNet(nn.Layer):
 
     def forward(self, x):
         x0 = paddle.shape(x)[0].reshape([1])
-        x1 = paddle.full([1], 1, dtype="int32")
+        x1 = paddle.full([1], 1, dtype="int64")
         out = paddle.concat([x0, x1])
         return out
